initial work on genome_features
